deployment/2-lambda-and-codebuild-executors/lambda/iac-ci-executors/requirements.txt
awscli==1.32.20
botocore==1.34.20
colorama==0.4.4
docutils==0.16
jmespath==1.0.1
pyasn1==0.5.1
python-dateutil==2.8.2
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.10.0
semantic-version==2.10.0
setuptools-rust==1.8.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.18
python-dotenv==0.21.1
cryptography==3.3.2
requests==2.31.0